requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
feedparser==5.2.1
idna==2.9
Pillow==7.1.2
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
